+++ crates/plugin_static_assets/src/lib.rs
@@ -1,18 +1,21 @@
 #![feature(path_file_prefix)]
 
-use std::path::Path;
+use std::{collections::HashMap, path::Path};
 
 use base64::engine::{general_purpose, Engine};
 use farmfe_core::{
   config::Config,
-  module::ModuleType,
-  plugin::Plugin,
-  resource::{Resource, ResourceType},
+  module::{ModuleId, ModuleType},
+  plugin::{
+    constants::PLUGIN_BUILD_STAGE_META_RESOLVE_KIND, Plugin, PluginHookContext, ResolveKind,
+  },
+  resource::{Resource, ResourceOrigin, ResourceType},
 };
 use farmfe_toolkit::{
   fs::{read_file_raw, read_file_utf8, transform_output_filename},
   lazy_static::lazy_static,
 };
+use farmfe_utils::stringify_query;
 
 // Default supported static assets: png, jpg, jpeg, gif, svg, webp, mp4, webm, wav, mp3, wma, m4a, aac, ico, ttf, woff, woff2
 lazy_static! {
@@ -21,6 +24,8 @@ lazy_static! {
     "ico", "ttf", "woff", "woff2",
   ];
 }
+
+const PLUGIN_NAME: &str = "FarmPluginStaticAssets";
 
 pub struct FarmPluginStaticAssets {}
 
@@ -32,20 +37,11 @@ impl FarmPluginStaticAssets {
 
 impl Plugin for FarmPluginStaticAssets {
   fn name(&self) -> &str {
-    "FarmPluginStaticAssets"
+    PLUGIN_NAME
   }
   /// Make sure this plugin is executed last
   fn priority(&self) -> i32 {
     99
-  }
-
-  fn resolve(
-    &self,
-    _param: &farmfe_core::plugin::PluginResolveHookParam,
-    _context: &std::sync::Arc<farmfe_core::context::CompilationContext>,
-    _hook_context: &farmfe_core::plugin::PluginHookContext,
-  ) -> farmfe_core::error::Result<Option<farmfe_core::plugin::PluginResolveHookResult>> {
-    Ok(None)
   }
 
   fn load(
@@ -84,12 +80,20 @@ impl Plugin for FarmPluginStaticAssets {
     context: &std::sync::Arc<farmfe_core::context::CompilationContext>,
   ) -> farmfe_core::error::Result<Option<farmfe_core::plugin::PluginTransformHookResult>> {
     if matches!(param.module_type, ModuleType::Asset) {
+      let resolve_kind = ResolveKind::from(
+        param
+          .meta
+          .get(PLUGIN_BUILD_STAGE_META_RESOLVE_KIND)
+          .unwrap()
+          .as_str(),
+      );
+
       if param.query.iter().find(|(k, _)| k == "inline").is_some() {
         let file_raw = read_file_raw(param.resolved_path)?;
         let file_base64 = general_purpose::STANDARD.encode(&file_raw);
         let path = Path::new(param.resolved_path);
         let ext = path.extension().and_then(|s| s.to_str()).unwrap();
-
+        // TODO: recognize MIME type
         let content = format!(
           "export default \"data:image/{};base64,{}\"",
           ext, file_base64
@@ -136,7 +140,11 @@ impl Plugin for FarmPluginStaticAssets {
             bytes,
             emitted: false,
             resource_type: ResourceType::Asset(ext.to_string()),
-            resource_pot: "STATIC_ASSETS".into(),
+            origin: ResourceOrigin::Module(ModuleId::new(
+              param.resolved_path,
+              &stringify_query(&param.query),
+              &context.config.root,
+            )),
             preserve_name: false,
           },
         );