+++ examples/script-entry/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @farmfe-examples/script-entry
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [d8eeda9]
+  - @farmfe/core@0.8.7
+  - @farmfe/cli@0.5.3
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [8a32a53]
+  - @farmfe/core@0.8.6
+  - @farmfe/cli@0.5.3
+
 ## 0.0.26
 
 ### Patch Changes